Pin `terraform-aws-rds-aurora` example test
Pin to the commit associated with v9.0.2 of the module. The most recent release (v9.1.0) uses new attributes on resource/aws_rds_cluster (domain and domain_iam_role_name) which were included in v5.37.0 of the TF AWS provider, but the Pulumi AWS provider has not yet been updated to use this version of the TF AWS provider.