Remove unnecessary mangling causing TF bridge issues.
Right now we mangle all locals and vars and loop variables to keep
things consistent with PCL style, but this is causing problems in
converting to PCL when there are other resources from openTofu that must be
bridged.  One such example is #181, where a local loop rename is
inconsistent with a locally renamed config var.

Surgical renaming when we should and not fixes this issue.  Among this
issue are a few other hidden naming bugs that only happened to work by
coincidental name alignment, and other unknowns.  It isn't clear which
bugs will be fixed by this (Or if it will introduce others!!!) so
afterwards a careful audit of open issues will be necessary and close
monitoring.

This unblocks #186 and may fix a number of other name mangling issues.

Fixes: #181