Add package block generation to tf to pcl conversion.
This is needed to implement automatic import of TF providers on
conversion.  This takes the work from pulumi/pulumi#17589 and adds it
into converted tf -> pp files.

After this, any sdk can leverage this information in it's conversions
form pcl to parameterize (and include without "detecting" it in the code
the given provider (usually pulumi-terraform-provider, the dynamic
bridge, to generate local sdks for tf providers).

Related #186