Test list and map objects as attributes
List and map objects can be assigned as attributes or blocks. We had
only really been testing them as blocks, both in the complex_X tests and
also the list_as_blocks and maps_as_blocks tests.

This changes the complex_X tests to instead use attribute syntax, to
test that works.