Require v0.6.0 of the `terraform-provider` instead of v0.5.4 because the latter is no longer available (#259)
In pulumi/pulumi, the test `TestConvertTerraformProviderGo` is failing.
It is an integration test that tries to convert a terraform program that
uses a dynamically bridged provider (for `supabase`). The reason is
that this terraform converter specifies `terraform-provider` v0.5.4
which by looking at the
[releases](https://github.com/pulumi/pulumi-terraform-provider/releases),
no longer exists. Instead in this PR using latest v0.6.0

This PR also applied formatting for the `tf.go` file