Bump the test schema for `pulumi-std` (#279)
This repository uses a vendored copy of the schema for the `pulumi-std`
provider in its tests. This change bumps this schema to the latest
released copy (2.1.0), ahead of changing the converter to use some new
functionality in this version. It also tweaks the linked version of
`pulumi/terraform`, since it appears that the referenced v1.4.0 doesn't
exist as a tag in that repository.