Clean up formatting and linting
As pointed out by @daveset in #296, linting hasn't been working on this
repository for a while, and formatting is a bit all over the place. This
change turns linting back on, removes deprecated linters and copies the
"big set" we use in `pulumi/pulumi`. With that we can clean it all up,
turn CI checks back on and deprecate a few other bits (e.g. linting now
takes care of copyright checks for us).