Support converting remote modules to resources parameterized by terraform-module (#324)
### Descriptions

This PR implements support for converting module calls into
(parameterized) resources backed by the `pulumi-terraform-module`
provider when annotated with `// @sandbox <packageName>` this way, we no
longer to download and convert the module sources, instead we convert
the module call into a resource block in PCL and add a package block
with `package <packageName> { ... }` that includes the parameterization
details of the sandboxed module.

When a module call is annotated with `// @sandbox <packageName>`, other
module calls with the same source will be considered annotated/sandboxed
with the same `<packageName>` such that the users do need to repeat the
annotation.

See `sandbox_multiple_registry_modules` test for examples.

~~The test is marked `partial` because the test package loader doesn't
handle loading parameterized packages~~

EDIT: the test is no longer partial and it type-checks

### Implementation details

We keep track of the module calls that are annotated with `// @sandbox
<packageName>` in the conversion state and emit the associated package
blocks at the top level packages PCL file. Before we translate module
calls, we check if they are sandboxed, in which case we emit a resource
block instead of downloading the module sources.

When converting inputs of sandboxed modules, we maintain their casing,
as per the convention of `pulumi-terraform-module` so we don't camelCase
the input fields of these modules. Similarly, when referencing outputs
of sandboxed modules, we keep the casing the output fields.

Fixes #323