Don't copy TF files to the output directory (#339)
It seems to be cleared not to copy TF files into the Pulumi output
directory. All the other files are still copied by default because they
may be required for a successful deployment.

Surprisingly, I couldn't find a standard function to do that, so I
drafted my own.

Also, adjusted the Claude's prompt and input arguments to make sure it
reads files from the source directory but writes to the output one. The
tests passed for me locally.