Extract the Claude prompt from the Pulumi MCP server (#342)
The Pulumi MCP server now contains a prompt that helps MCP hosts convert
Terraform to Pulumi programs. This PR switches the Claude benchmark from
a simplistic hard-coded prompt to the one from MCP.

One complication is that currently there is no way to tell Claude Code
to use a given MCP prompt directly. Therefore, this PR goes in a
roundabout way: it calls an MCP client to retrieve the prompt
(parameterized with the folder name) and then passes this prompt to
Claude Code.