Handle implicitly required (non-Pulumi) providers as parameterized ones (#378)
### Description

When we added support for Any TF provider in the converter, we would
only emit parameterized package blocks for TF providers that are
specified in `required_providers` block in the Terraform program.
However, if a user doesn't specify a provider in `required_providers`
and we don't have it bridged then we result in issues such as the one
described in #371

This PR fixes the problem by resolving implicitly required providers as
parameterized ones as if they were in the `required_providers` block and
resolving their version to latest (if it exists).

The implementation adds `ProviderInfoResolver` interface to the
translation logic such that during tests, we handle test providers
(`simple`, `complex`, `renamed` etc.) differently from what we would do
to real providers since those are not specified either in a
`required_providers` block. In the main user code path, we use a default
implementation of `ProviderInfoResolver` that goes out to the registry
to retrieve the latest provider version.

Fixes #371