Set the tfe version during tests (#385)
`TestResolveLatestProviderVersion` and `TestTranslate` depended on a
outgoing call to a registry to fetch the latest and greatest tfe
Terraform module version. This change modifies the tests not to obtain
the last version, but to relay on a pre-defined version that should
never change.

This should reduce test maintenance by not having to update the tests
each time a new version of `tfe` is released while minimally sacrificing
test coverage (we are no longer calling the actual APIs to obtain what
the latest version is for tfe).

Fixes #384 
Fixes #383